requirements.txt
pandas~=2.2.3
aiogram~=3.19.0
pillow~=11.2.1
aiogram_dialog~=2.3.1
numpy~=2.2.6
regex~=2024.11.6
pytest~=8.3.5
langchain~=0.3.24
tqdm~=4.67.1
pymorphy3~=2.0.3
python-dateutil~=2.9.0.post0
reportlab~=4.4.1
openpyxl
sentence_transformers
langchain_community
aiogram_media_group
langchain_gigachat
faiss-cpu
